requirements.txt
cachetools==2.0.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
configurate==0.4.1
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
MarkupSafe==1.0
PyYAML==3.12
requests==2.18.4
urllib3==1.22
Werkzeug==0.12.2
